PB-272: Adding Mocha JS Coverage Command and Viewer
Added the solum:test:mocha:coverage symfony command to make running
test suites simpler.  Because it is a little bit complicated to make
unit testing and code coverage work with mocha, I have created standards
where I expect certain index files to be there before the tests will be
considered for coverage.

Specifically, the structure must be:
 - [js suite dir]
   - lib
     - [Your source code here]
   - test
     - mocha
       - runSuite.js - executes the suite definition lib with the non-coverage lib
       - runSuiteCoverage.js - executes the suite with the instrumented "lib-cov" source
       - suiteDefinitions.js - Uses node's "require" to gather and configure all mocha
                               tests in this suite and passes them which lib to use.
       - [All of your tests in the appropriate directories]

The interface will look in the [project root]/web/coverage/mocha/ directory to find
previously generated coverage reports.

I tried to run these reports via a command on a page but it cannot be run as
apache:apache because of all of the copying and environment/node differences.

You must use the command line to generate the file before going to the interface.