PB-272: Create the SolumBundle: Refactored "Model" to "Component"
After discussions with Rob, the term "model" is too confusing for the
built in reusable objects within solum, such as the page object or
the grouped list object.  In order to avoid this confusion, the solum
library will be said to have "components" which can be used in your
page "models".